move ProgramSummaries into model
