+++ explore/src/main/scala/explore/observationtree/AsterismGroupObsListActions.scala
@@ -35,57 +35,6 @@ import scala.annotation.unused
 import scala.collection.immutable.SortedSet
 
 object AsterismGroupObsListActions {
-  // private def obsDropGetter(
-  //   draggedIds: ObsIdSet
-  // ): ProgramSummaries => Option[AsterismGroup] =
-  //   _.asterismGroups.findContainingObsIds(draggedIds)
-
-  // private def obsDropSetter(draggedIds: ObsIdSet, srcIds: ObsIdSet, destIds: ObsIdSet)(
-  //   oAsterismGroup: Option[AsterismGroup]
-  // ): ProgramSummaries => ProgramSummaries = ps => {
-  //   val origAsterismGroups = ps.asterismGroups
-
-  //   // should always have an asterism group and be able to find the dragged Ids in the list
-  //   (oAsterismGroup, origAsterismGroups.findContainingObsIds(draggedIds))
-  //     .mapN { case (destGroup, srcGroup) =>
-  //       // no matter what, the src group goes away because all or some of the obs were moved.
-  //       val tempList    = origAsterismGroups - srcGroup.obsIds
-  //       // if we didn't move all of the obs, we need to add the remnant back into the list
-  //       val updatedList =
-  //         srcGroup.removeObsIds(draggedIds).fold(tempList) { filteredGroup =>
-  //           tempList + filteredGroup.asObsKeyValue
-  //         }
-
-  //       val updatedGroupList = if (srcGroup.obsIds === srcIds) {
-  //         // do/redo - we have to remove the dest and add it back with the extra observations
-  //         // what is really in the destination is a group with the destIds
-  //         updatedList - destIds + AsterismGroup(destIds ++ draggedIds,
-  //                                               destGroup.targetIds
-  //         ).asObsKeyValue
-  //       } else if (destGroup.obsIds === srcIds) {
-  //         // undo - we have to put the destination group back and, if it exists, remove the
-  //         // group with (srcIds - draggedIds)
-  //         val tmp = updatedList + destGroup.asObsKeyValue
-  //         srcIds.remove(draggedIds).fold(tmp)(ids => tmp - ids)
-  //       } else updatedList // something changed outside the scope of this undoctx
-
-  //       // remove the dragged ids from all the targets in the src
-  //       val tempTargets = srcGroup.targetIds.foldRight(ps.targetsWithObs) {
-  //         case (tid, targetWithObs) =>
-  //           targetWithObs.updatedWith(tid)(_.map(_.removeObsIds(draggedIds)))
-  //       }
-
-  //       // add the dragged ids to all the targets in the destination
-  //       val updatedTargetsWithObs = destGroup.targetIds.foldRight(tempTargets) {
-  //         case (tid, targetWithObs) =>
-  //           targetWithObs.updatedWith(tid)(_.map(_.addObsIds(draggedIds)))
-  //       }
-
-  //       ps.copy(asterismGroups = updatedGroupList, targetsWithObs = updatedTargetsWithObs)
-  //     }
-  //     .getOrElse(ps)
-  // }
-
   private def updateExpandedIds(
     draggedIds: ObsIdSet,
     optDestIds: Option[ObsIdSet]
@@ -117,21 +66,10 @@ object AsterismGroupObsListActions {
         .andThen(KeyedIndexedList.value)
         .andThen(ObsSummary.scienceTargetIds)
 
-    // undoCtx
-    //   // This deserves an explanation:
-    //   // The traversal provides a View over observations that have the same
-    //   // Asterism. Therefore, we can see this as a view over a single
-    //   // ConstraintSet. We get the value from any of them (eg: head), and
-    //   // we set it in all of them.
-    //   .zoom(traversal.getAll.andThen(_.head), traversal.modify)
-
-    // Action(getter = obsDropGetter(draggedIds), setter = obsDropSetter(draggedIds, srcIds, destIds))(
     Action(getter = traversal.getAll.andThen(_.head), setter = traversal.replace)(
       onSet = (observationList, asterismIds) =>
-        // oAsterismGroup.foldMap { asterismGroup =>
         // destination ids may not be found when undoing
         val filteredTargetIds = asterismIds.filter(allTargets.contains)
-        // agwo.asterismGroups.findWithTargetIds(asterismGroup.targetIds).map(_.obsIds)
         AsterismQueries.replaceAsterism[IO](
           programId,
           draggedIds.toList,
