+++ explore/src/main/scala/explore/targets/TargetAddDeleteActions.scala
@@ -10,9 +10,9 @@ import clue.FetchClient
 import clue.data.syntax.*
 import explore.DefaultErrorPolicy
 import explore.common.AsterismQueries
-import explore.common.AsterismQueries.*
 import explore.model.AsterismGroup
 import explore.model.ObsIdSet
+import explore.model.ProgramSummaries
 import explore.model.TargetWithObs
 import explore.model.syntax.all.*
 import explore.undo.*
