+++ model-testkit/shared/src/main/scala/explore/model/arb/package.scala
@@ -3,7 +3,6 @@
 
 package explore.model.arb
 
-import explore.model.arb.ArbAsterism
 object all
     extends ArbTargetVisualOptions
     with CogenUUID
