+++ model/shared/src/main/scala/explore/model/syntax/package.scala
@@ -4,15 +4,20 @@
 package explore.model.syntax
 
 import cats.syntax.all.*
+import explore.model.AsterismGroup
+import explore.model.AsterismGroupList
+import explore.model.ObsIdSet
 import explore.model.enums.PosAngleOptions
 import lucuma.core.enums.Site
 import lucuma.core.math.Angle
 import lucuma.core.math.Declination
 import lucuma.core.model.ConstraintSet
 import lucuma.core.model.PosAngleConstraint
+import lucuma.core.model.Target
 import lucuma.core.util.TimeSpan
 
 import scala.collection.immutable.SortedMap
+import scala.collection.immutable.SortedSet
 
 object all:
 
@@ -31,6 +36,13 @@ object all:
       case PosAngleConstraint.AverageParallactic     => PosAngleOptions.AverageParallactic
       case PosAngleConstraint.ParallacticOverride(_) => PosAngleOptions.ParallacticOverride
       case PosAngleConstraint.Unbounded              => PosAngleOptions.Unconstrained
+
+  extension (self: AsterismGroupList)
+    def findContainingObsIds(obsIds: ObsIdSet): Option[AsterismGroup] =
+      self.find { case (ids, _) => obsIds.subsetOf(ids) }.map(AsterismGroup.fromTuple)
+
+    def findWithTargetIds(targetIds: SortedSet[Target.Id]): Option[AsterismGroup] =
+      self.find { case (_, grpIds) => grpIds === targetIds }.map(AsterismGroup.fromTuple)
 
   extension [A](list: Iterable[A])
     def toSortedMap[K: Ordering, V](getKey: A => K, getValue: A => V = identity[A](_)) =
